refactor: command separated
