+++ src/services/config/mod.rs
@@ -1,2 +1,4 @@
+mod command;
 pub mod config_actor;
 mod constructor;
+pub(super) use command::ConfigCommand;
